Support for XTB seems to work
I added support from S. Grimme's XTB program, so one can use GFN2-xtb
in QMMM calculations with pDynamo. It passes preliminary tests, but there
are things that could be wrong. Also, I am not sure if the electrostatic embedding
is "production ready" in xtb.

I also made small changes in the TM interface, so now the deletion of the gradient
file happens in the translator, not in the fakeorca script.