The tests seem to work correctly now. Use of the translator with XTB is probably as safe as it gets.
I think this is as well-tested as we will get, since we have nothing to comare to.
The translator works correctly when there are no point charges (pure QM calculations), as
as the RMSD after optimizinc crystal structures is similar to what you get with other methods.
It also seems to work correctly with charges also, but that is harder to check accurately.