Single points now supported, and improvements to the README
I added a "-SP" flag to OrcaTranslator.py so it handles single-point calculations in TM
(xtb is so fast that I don't think it's worth it the extra code to skip the gradients, at
least for now).

I also updated the README with instructions on how to set up calculations with TM
and xtb.